refactor(#837): catalog CA skeleton — domain/itinerary, clustering, geo
Closes #837
Parent #829

Move pure plan/cluster/geo kernels into src/domain; leave application/adapters
skeletons; TODO markers on api/* pending vertical slices. Written by opencode
(opencode-go/deepseek-v4-flash, variant max).
